Better garbage collection
Previously each task had a reference to other tasks in its plan, causing
memory for all tasks to be held as long as a reference to one task was
strongly reachable. This change uses indirection to reduce the amount of
memory reachable: we uniquely identify each task in a plan and use this
for building relationships for trace purposes.